Resolve TODO
